75-78 bug fixes

+++ app/src/angular/controllers/guest/create/step-5.controller.js
@@ -11,6 +11,10 @@ function GuestKeystoreCreateStep5Controller($rootScope, $scope, $log, $state, $s
 
     $scope.importKycFile = (event) => {
         RPCService.makeCall('importKYCPackage', { walletId: $rootScope.wallet.id }).then((walletSetting) => {
+            //on cancel choose a file
+            if (!walletSetting) {
+                return;
+            }
             SelfkeyService.triggerAirdrop(walletSetting.airDropCode).then(()=>{
                 SqlLiteService.removeAirdropCode();
             });
